backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.5
httpx==0.25.2
redis==5.0.1
celery==5.3.4
pydantic[email]==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1

# AI and ML
transformers==4.36.0
torch==2.1.1
numpy==1.25.2
pillow==10.1.0

# News and fact-checking APIs
newsapi-python==0.2.7
googlesearch-python==1.2.3

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1